Change how the game URI is generated
This commit updates how the game URI is calculated. Initially, I
took an approach that would allow me to easily check whether a
game solution is valid by hashing the expected solution and using
that as the URI. This meant any solution could be easily validated
by hashing it and checking for eqality.

The problem with this is that a solution can only ever be seen once,
otherwise our UIDs aren't, well, unique. Instead, we can embed the
solution as part of the game struct, and strip it out when returning
the data to the user.