Formalize choices into its own struct, distinct from employee headshot
While it would probably be ok to just return the entire employee
headshot, it does give away some information that we may not want
to give to the user (like the alt image text) and it contains some
irrelevant information like the image ID that our game doesn't
care about.

This commit changes the new game function to use the choices struct
rather than the employee headshot. None of the downstream details
are affected, since this is in our data layer.