Fix bug where single game details were being shown as array
This commit will fix a bug where our helper function was returning
an array regardless of how many games were present. This caused
conflicts when we were trying to strip solutions out of a single
game and did -not- want an array. Now, the helper function is
split into two to distinguish between what return type is desired.